[Refactoring] Repository 패키지 생성
