+++ src/main/java/aresain/loldatastats/loldata/participant/ParticipantService.java
@@ -12,6 +12,9 @@ import aresain.loldatastats.entity.ParticipantAnalysis;
 import aresain.loldatastats.entity.ParticipantPerk;
 import aresain.loldatastats.entity.ParticipantSummary;
 import aresain.loldatastats.loldata.participant.dto.ParticipantSummaryDto;
+import aresain.loldatastats.loldata.participant.repository.ParticipantAnalysisRepository;
+import aresain.loldatastats.loldata.participant.repository.ParticipantPerkRepository;
+import aresain.loldatastats.loldata.participant.repository.ParticipantSummaryRepository;
 import aresain.loldatastats.riot.dto.match.ParticipantDto;
 import aresain.loldatastats.riot.dto.match.PerkStyleDto;
 import aresain.loldatastats.riot.dto.match.PerksDto;
